requirements.txt
python-nmap==0.6.1
slackclient==2.2.1
PyYAML==5.1.2